backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
earthengine-api==0.1.384
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
openai==1.3.5
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
geopy==2.4.0
Pillow==10.1.0
httpx==0.25.2
